refactor: cars 클래스 분리
