+++ src/test/java/racingcar/GameManagerTest.java
@@ -31,50 +31,4 @@ public class GameManagerTest {
     }
 
 
-    /*
-     * 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4 이상일 경우 전진하고, 3 이하의 값이면 멈춘다.
-     * */
-    @Test
-    public void testForward() {
-
-        Car car = new Car("Test");
-        car.goForward(4);
-        Assertions.assertThat(car.getPosition()).isEqualTo(1);
-        car.goForward(3);
-        Assertions.assertThat(car.getPosition()).isEqualTo(1);
-        car.goForward(9);
-        Assertions.assertThat(car.getPosition()).isEqualTo(2);
-    }
-
-    @Test
-    public void testSinglewinner() {
-
-        List<Car> cars = gameManager.makeCarList(Arrays.asList("pobi", "crong", "honux", "teo"));
-        int[] fixedNumber = {1, 2, 3, 6};
-        for (int i = 0; i < 12; i++) {
-            cars.get(i % 4).goForward(fixedNumber[i % 4]);
-        }
-
-        List<Car> winners = Arrays.asList(new Car("teo", 3));
-        List<Car> rank = gameManager.getWinners(cars);
-        for (int i = 0; i < winners.size(); i++) {
-            Assertions.assertThat(winners.get(i).equals(rank.get(i))).isEqualTo(true);
-        }
-
-    }
-
-    @Test
-    public void testMultiWinner() {
-
-        List<Car> cars = gameManager.makeCarList(Arrays.asList("pobi", "crong", "honux", "teo"));
-        int[] fixedNumber = {5, 2, 3, 6};
-        for (int i = 0; i < 12; i++) {
-            cars.get(i % 4).goForward(fixedNumber[i % 4]);
-        }
-        List<Car> winners = Arrays.asList(new Car("pobi", 3), new Car("teo", 3));
-        List<Car> rank = gameManager.getWinners(cars);
-        for (int i = 0; i < winners.size(); i++) {
-            Assertions.assertThat(winners.get(i).equals(rank.get(i))).isEqualTo(true);
-        }
-    }
 }
